Beta branch

+++ frontend/src/contexts/UserContext.tsx
@@ -1,12 +1,18 @@
 import React, { createContext, useState, useContext, useEffect } from "react";
 import { onAuthStateChanged } from "firebase/auth";
 import { auth } from "../services/firebase";
+import { doc, getDoc } from "firebase/firestore";
+import { db } from "../services/firebase";
 
 interface User {
-  displayName: string | null;
+  firebase_uid: string;
+  username: string | null;
   email: string | null;
-  photoURL: string | null;
-  uid: string | null;
+  display_picture: string | null;
+  full_name: string | null;
+  email_verified: boolean;
+  isSSO: boolean;
+  account_type: 'free' | 'premium';
 }
 
 interface UserContextProps {
@@ -25,22 +31,23 @@ export const UserProvider: React.FC<{ children: React.ReactNode }> = ({
   });
 
   useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (firebaseUser) => {
+    const unsubscribe = onAuthStateChanged(auth, async (firebaseUser) => {
       if (firebaseUser) {
-        const userData: User = {
-          displayName: firebaseUser.displayName,
-          email: firebaseUser.email,
-          photoURL: firebaseUser.photoURL,
-          uid: firebaseUser.uid,
-        };
-        setUser(userData);
-        localStorage.setItem("userData", JSON.stringify(userData));
+        const token = await firebaseUser.getIdToken();
+        const userDoc = await getDoc(doc(db, "users", firebaseUser.uid));
+        if (userDoc.exists()) {
+          const userData = userDoc.data() as User;
+          setUser(userData);
+          localStorage.setItem("userToken", token); // Store token
+          localStorage.setItem("userData", JSON.stringify(userData));
+        }
       } else {
         setUser(null);
+        localStorage.removeItem("userToken"); // Remove token
         localStorage.removeItem("userData");
       }
     });
-
+  
     return () => unsubscribe();
   }, []);
